refactor: 인터페이스 삭제 후 추상 클래스를 Player 로 변경
